Added WasmJS implementation
