+++ composeApp/src/wasmJsMain/kotlin/org/noiseplanet/noisecapture/services/audio/JSAudioRecordingService.kt
@@ -3,18 +3,17 @@ package org.noiseplanet.noisecapture.services.audio
 import kotlinx.browser.window
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.await
 import kotlinx.coroutines.launch
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import org.noiseplanet.noisecapture.interop.MediaRecorder
 import org.noiseplanet.noisecapture.log.Logger
-import org.noiseplanet.noisecapture.util.OPFSHelper
+import org.noiseplanet.noisecapture.services.storage.FileSystemService
+import org.noiseplanet.noisecapture.services.storage.OPFSFileSystemService
 import org.noiseplanet.noisecapture.util.injectLogger
 import org.w3c.dom.mediacapture.MediaStream
 import org.w3c.dom.mediacapture.MediaStreamConstraints
 import org.w3c.files.Blob
-import org.w3c.files.File
 
 @OptIn(ExperimentalWasmJsInterop::class)
 class JSAudioRecordingService : AudioRecordingService, KoinComponent {
@@ -22,7 +21,7 @@ class JSAudioRecordingService : AudioRecordingService, KoinComponent {
     // - Properties
 
     private val logger: Logger by injectLogger()
-    private val audioStorageService: AudioStorageService by inject()
+    private val fileSystemService: FileSystemService by inject()
 
     private val scope = CoroutineScope(Dispatchers.Default)
     private var mediaRecorder: MediaRecorder? = null
@@ -58,20 +57,6 @@ class JSAudioRecordingService : AudioRecordingService, KoinComponent {
         mediaRecorder?.stop()
     }
 
-    override suspend fun getFileSize(audioUrl: String): Long? {
-        val (fileHandle, _) = OPFSHelper.getFileHandle(audioUrl) ?: return null
-        val file: File = fileHandle.getFile().await()
-
-        return file.size.toInt().toLong()
-    }
-
-    override fun deleteFileAtUrl(audioUrl: String) {
-        scope.launch {
-            val (fileHandle, directoryHandle) = OPFSHelper.getFileHandle(audioUrl) ?: return@launch
-            directoryHandle.removeEntry(fileHandle.name).await()
-        }
-    }
-
 
     // - Private functions
 
@@ -87,7 +72,7 @@ class JSAudioRecordingService : AudioRecordingService, KoinComponent {
             blob?.let {
                 val url = "measurement/audio/$fileName.ogg"
                 scope.launch {
-                    audioStorageService.store(key = url, blob = it)
+                    (fileSystemService as? OPFSFileSystemService)?.store(key = url, blob = it)
                     recordingStopListener?.onRecordingStop(url)
                 }
             } ?: logger.warning("Could not get recorder audio URL: Blob was null.")
